Add files via upload
previous solution relied on strictly ordered sequence of bids (by construction).
current one tackles this. plus few bugs solved 
a. the first item was not added to disjoinedResult map. if it happened to have highest weight - this would not be found. It's solved here by iteration over the disjoinedResult map (not result map as it was a case earlier). We fill here disjoinedResult just identically like we did it before with result map
b. type of result map changed (to map string to Item object). result map is kept for decorative purposes (to reflect on initial bids)
c. redundant check was done in earlier solution (the newly arriving items set was checked with each and every sets in results map. if the disjointness answer was negative, the else() clause of that if() would check if such an element does not exist in if(disjointResult.count(coreKey) == 0) and add it to  disjointResult. Problem is the items would still be checked over and over against next subset in result map pointed by iterator. But there is no need anymore to add the items with that coreKey to the disjointResult. Solution was to get rid of else() and make an extra if() in the outer loop just below result initialization 

By doing points a), c) the following achieved: the items not included in disjointResults set are not considered in future. Hence the optimization of iteration is achieved.
Details on example are in last_remarks_Jan29